requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
pytz
django-redis
google-generativeai


